# src/docs/swagger-spec.yaml
openapi: 3.0.4
paths:
  /items:
    get:
      deprecated: false
      operationId: ItemsController_findAll
      parameters: []
      responses:
        '200':
          description: 正常系
          content:
            fetch:
              example:
                - id: '0000000000000'
                  name: testProduct1
                  price: 11111
                  description: テストディスクリプション
                  status: ON_SALE
                  userId: '0000000000000'
                - id: '0000000000000'
                  name: testProduct1
                  price: 11111
                  description: テストディスクリプション
                  status: SOLD_OUT
                  userId: '0000000000000'
      summary: itemsの一覧の取得
      tags:
        - items
    post:
      deprecated: false
      operationId: ItemsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: アイテム名
                  example: MacBook Pro 14inch
                  maxLength: 40
                price:
                  type: number
                  description: アイテムの価格（円）
                  example: 298000
                  minimum: 1
                description:
                  type: string
                  description: アイテムの詳細説明
                  example: Apple M2 Pro チップ搭載
                  maxLength: 1000
                  nullable: true
              required:
                - name
                - price
      responses:
        '201':
          description: 正常系
      security:
        - jwt: []
      summary: itemsの作成
      tags:
        - items
  /items/{id}:
    get:
      operationId: ItemsController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - Items
        - items
    put:
      operationId: ItemsController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: &ref_0
        - Items
    delete:
      operationId: ItemsController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
  /items/update/{id}:
    put:
      operationId: ItemsController_updatePrice
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemDto'
      responses:
        '200':
          description: ''
      tags: *ref_0
  /auth/signup:
    post:
      operationId: AuthController_signUp
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: ''
      tags:
        - Auth
  /auth/signin:
    post:
      deprecated: false
      operationId: AuthController_signIn
      parameters: []
      requestBody:
        required: true
        description: ログイン認証情報
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: ユーザーのメールアドレス
                  example: example@test.com
                password:
                  type: string
                  description: パスワード（8文字以上、大文字・小文字・数字・記号を各1つ以上含む）
                  example: Gtn-804918360
                  minLength: 8
              required:
                - email
                - password
            examples:
              validExample:
                summary: 有効な認証情報
                value:
                  email: user@example.com
                  password: StrongPass123!
      responses:
        '201':
          description: 正常系
      summary: ログイン
      tags:
        - auth
info:
  title: 練習WebAPI
  description: 練習WebAPIの説明欄
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    jwt:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CreateItemDto:
      type: object
      properties:
        name:
          type: string
          description: アイテム名
          example: MacBook Pro 14inch
          maxLength: 40
      required:
        - name
    CreateUserDto:
      type: object
      properties: {}